+++ src/integrators/integrate.rs
@@ -2,7 +2,7 @@ use ndarray::{ArrayView, Axis, Ix1};
 
 /// Implements Simpon's Rule integration on an array, representing sampled
 /// points of a periodic function.
-pub fn periodic_simpson_integrate(samples: ArrayView<f64, Ix1>, h: f64) -> f64 {
+pub fn periodic_simpson_integrate(samples: ArrayView<Float, Ix1>, h: Float) -> Float {
     let len = samples.dim();
 
     assert!(
@@ -26,6 +26,6 @@ pub fn periodic_simpson_integrate(samples: ArrayView<f64, Ix1>, h: f64) -> f64 {
 }
 
 /// Implements most straight forward step-sum integration method
-pub fn integrate(samples: ArrayView<f64, Ix1>, h: f64) -> f64 {
+pub fn integrate(samples: ArrayView<Float, Ix1>, h: Float) -> Float {
     (&samples * h).sum(Axis(0))[()]
 }
