+++ src/output.rs
@@ -1,6 +1,7 @@
 use crate::distribution::Distribution;
 use crate::flowfield::FlowField3D;
 use crate::particle::Particle;
+use crate::Float;
 use ndarray::{Array, Ix4};
 
 /// Captures values that can be outputed during simulation.
@@ -9,7 +10,7 @@ use ndarray::{Array, Ix4};
 pub struct OutputEntry {
     pub distribution: Option<Distribution>,
     pub flowfield: Option<FlowField3D>,
-    pub magneticfield: Option<Array<f64, Ix4>>,
+    pub magneticfield: Option<Array<Float, Ix4>>,
     pub particles: Option<Vec<Particle>>,
     pub timestep: usize,
 }
